-Wmissing-attributes: avoid duplicates and false positives
The initial patch for PR 81824 fixed one of the possibilities of
-Wmissing-attributes reporting duplicates, namely, if TMPL had an
attribute in ATTRLIST that was missing from DECL's decl and type
attribute lists, both being non-empty.

Another possibility of duplicate reporting remained: when an attribute
in ATTRLIST is present in both decl and type attribute lists of TMPL,
and absent from DECL's attribute lists, it is reported once for each
of TMPL's lists.

The implementation still allowed for false positives: an attribute in
ATTRLIST that is present in TMPL will be regarded as missing as soon
as it is not found in DECL's decl attribute list, even if it is later
found in DECL's type attribute list.

This patch fixes both problems, so that an attribute from ATTRLIST
that is present in any of TMPL's lists will be reported at most once,
and only if it is missing from both DECL's lists.


Now, I realize there are some attributes that are only acceptable for
types, and some that are only acceptable for function declarations.
ISTM that some even have different meanings depending on whether
they're associated with types or declarations.  There's room for an
argument for checking only corresponding lists, for at least some of
the attributes.  AFAICT it doesn't apply to -Wmissing-attributes, that
are either acceptable in either list, or only in the FUNCTION_DECL
list, so I'm leaving it at that in the hope that it doesn't apply to
any other users of decls_mismatched_attributes either.


for  gcc/ChangeLog

	PR middle-end/81824
	* attribs.c (decls_mismatched_attributes): Avoid duplicates
	and false positives.

for  gcc/testsuite/ChangeLog

	PR middle-end/81824
	* g++.dg/Wmissing-attributes-1.C: New.